+++ ext/types/args.go
@@ -1,5 +1,3 @@
-//+build experimental
-
 package types
 
 //go:generate go run ./_codegen/args.go