+++ ext/types/args_gen.go
@@ -1,5 +1,3 @@
-//+build experimental
-
 package types
 
 // LookupString retrieves an argument of type string from the list stored under the specified the index.