+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Added
 
 - **errors:** Composite handler from Emperror
+- **errors:** Test handler from Emperror
 
 
 ## 0.11.0 - 2017-08-28